added photo gallery

+++ src/components/ViewKeebList/index.js
@@ -6,10 +6,11 @@ import API from '../../utils/API'
 // Contexts
 import { useProfile, useProfileData } from '../../contexts/ProfileContext'
 // Material-UI Components
-import {Avatar, Card, List, ListItem, ListItemAvatar, ListItemText, ListItemSecondaryAction, IconButton, } from '@material-ui/core'
+import { Avatar, Card, List, ListItem, ListItemAvatar, ListItemText, ListItemSecondaryAction, IconButton, Tooltip, Typography } from '@material-ui/core'
 // Material-UI Icons
 import KeyboardIcon from '@material-ui/icons/Keyboard';
 import RemoveCircleOutlineIcon from '@material-ui/icons/RemoveCircleOutline';
+import MusicNoteIcon from '@material-ui/icons/MusicNote';
 // Material-UI Styles
 import { makeStyles } from '@material-ui/core/styles'
 const useStyles = makeStyles({
@@ -21,11 +22,14 @@ const useStyles = makeStyles({
         dispaly: 'flex',
         flexDirection: "column",
         width: '100%',
-        height:"510px"
+        height: "510px"
     },
     keebList: {
         height: '400px',
         overflow: "auto"
+    },
+    keebIcon: {
+        color: 'white'
     }
 })
 
@@ -42,9 +46,13 @@ export default function ViewKeebList(props) {
     const handleRemoveKeeb = event => {
         event.preventDefault()
         let id = event.currentTarget.id
-        API.deleteKeeb(profile.token, id).then(res => {
-            window.location.reload()
-        })
+        let name = event.currentTarget.getAttribute('name')
+        let confirmAlert = window.confirm(`Are you sure to delete ${name}`)
+        if (confirmAlert === true) {
+            API.deleteKeeb(profile.token, id).then(res => {
+                window.location.reload()
+            })
+        }
     }
 
     return (
@@ -56,25 +64,37 @@ export default function ViewKeebList(props) {
                         <ListItem key={res.id} divider>
                             <ListItemAvatar>
                                 <Avatar>
-                                    <KeyboardIcon />
+                                    <Tooltip title={`Update ${res.maker} ${res.name} photos`} arrow>
+                                        <IconButton
+                                            className={classes.keebIcon}
+                                            onClick={() => window.location.href = `/updatekeeb/photos/${res.id}`}
+                                        >
+                                            <KeyboardIcon />
+                                        </IconButton>
+                                    </Tooltip>
                                 </Avatar>
                             </ListItemAvatar>
                             <ListItemText
                                 primary={
-                                    <Link to={`/updatekeeb/${res.id}`} className={classes.updateKeebLink}>
-                                        <strong>Update: {res.maker} {res.name}</strong>
-                                    </Link>
+                                    <Tooltip title={`Update ${res.maker} ${res.name} parts`} arrow>
+                                        <Link to={`/updatekeeb/${res.id}`} className={classes.updateKeebLink}>
+                                            <strong>Update: {res.maker} {res.name}</strong>
+                                        </Link>
+                                    </Tooltip>
                                 }
                             />
                             <ListItemSecondaryAction>
-                                <IconButton
-                                    edge="end"
-                                    aria-label="remove"
-                                    id={res.id}
-                                    onClick={handleRemoveKeeb}
-                                >
-                                    <RemoveCircleOutlineIcon />
-                                </IconButton>
+                                <Tooltip title={`Delete ${res.maker} ${res.name}`} arrow>
+                                    <IconButton
+                                        edge="end"
+                                        aria-label="remove"
+                                        name={`${res.maker} ${res.name}`}
+                                        id={res.id}
+                                        onClick={handleRemoveKeeb}
+                                    >
+                                        <RemoveCircleOutlineIcon />
+                                    </IconButton>
+                                </Tooltip>
                             </ListItemSecondaryAction>
                         </ListItem>
                     ))
@@ -82,6 +102,6 @@ export default function ViewKeebList(props) {
                     null
                 }
             </List>
-        </Card>
+        </Card >
     )
 }
